move rnntl out of prototype
