+++ test/torchaudio_unittest/functional/functional_impl.py
@@ -9,7 +9,18 @@ import torchaudio.functional as F
 from parameterized import parameterized
 from scipy import signal
 
-from torchaudio_unittest.common_utils import TestBaseMixin, get_sinusoid, nested_params, get_whitenoise
+from torchaudio_unittest.common_utils import (
+    TestBaseMixin,
+    get_sinusoid,
+    nested_params,
+    get_whitenoise,
+    compute_with_numpy_transducer,
+    compute_with_pytorch_transducer,
+    get_basic_data,
+    get_B1_T2_U3_D5_data,
+    get_B2_T4_U3_D3_data,
+    get_random_data,
+)
 
 
 class Functional(TestBaseMixin):
@@ -41,6 +52,15 @@ class Functional(TestBaseMixin):
         estimate = estimate[..., n_to_trim:-n_to_trim]
 
         self.assertEqual(estimate, ground_truth, atol=atol, rtol=rtol)
+
+    def _test_costs_and_gradients(
+        self, data, ref_costs, ref_gradients, atol=1e-6, rtol=1e-2
+    ):
+        logits_shape = data["logits"].shape
+        costs, gradients = compute_with_pytorch_transducer(data=data)
+        self.assertEqual(costs, ref_costs, atol=atol, rtol=rtol)
+        self.assertEqual(logits_shape, gradients.shape)
+        self.assertEqual(gradients, ref_gradients, atol=atol, rtol=rtol)
 
     def test_lfilter_simple(self):
         """
@@ -435,6 +455,44 @@ class Functional(TestBaseMixin):
         waveform = torch.rand(2, 44100 * 1, dtype=self.dtype, device=self.device)
         waveform_shift = F.pitch_shift(waveform, sample_rate, n_steps)
         assert waveform.size() == waveform_shift.size()
+
+    def test_rnnt_loss_basic_backward(self):
+        logits, targets, logit_lengths, target_lengths = get_basic_data(self.device)
+        loss = F.rnnt_loss(logits, targets, logit_lengths, target_lengths)
+        loss.backward()
+
+    def test_rnnt_loss_basic_forward_no_grad(self):
+        logits, targets, logit_lengths, target_lengths = get_basic_data(self.device)
+        logits.requires_grad_(False)
+        F.rnnt_loss(logits, targets, logit_lengths, target_lengths)
+
+    @parameterized.expand([
+        (get_B1_T2_U3_D5_data, torch.float32, 1e-6, 1e-2),
+        (get_B2_T4_U3_D3_data, torch.float32, 1e-6, 1e-2),
+        (get_B1_T2_U3_D5_data, torch.float16, 1e-3, 1e-2),
+        (get_B2_T4_U3_D3_data, torch.float16, 1e-3, 1e-2),
+    ])
+    def test_rnnt_loss_costs_and_gradients(self, data_func, dtype, atol, rtol):
+        data, ref_costs, ref_gradients = data_func(
+            dtype=dtype,
+            device=self.device,
+        )
+        self._test_costs_and_gradients(
+            data=data,
+            ref_costs=ref_costs,
+            ref_gradients=ref_gradients,
+            atol=atol,
+            rtol=rtol,
+        )
+
+    def test_rnnt_loss_costs_and_gradients_random_data_with_numpy_fp32(self):
+        seed = 777
+        for i in range(5):
+            data = get_random_data(dtype=torch.float32, device=self.device, seed=(seed + i))
+            ref_costs, ref_gradients = compute_with_numpy_transducer(data=data)
+            self._test_costs_and_gradients(
+                data=data, ref_costs=ref_costs, ref_gradients=ref_gradients
+            )
 
 
 class FunctionalCPUOnly(TestBaseMixin):
