+++ test/torchaudio_unittest/functional/torchscript_consistency_cpu_test.py
@@ -1,10 +1,10 @@
 import torch
 
 from torchaudio_unittest.common_utils import PytorchTestCase
-from .torchscript_consistency_impl import Functional
+from .torchscript_consistency_impl import Functional, FunctionalFloat32Only
 
 
-class TestFunctionalFloat32(Functional, PytorchTestCase):
+class TestFunctionalFloat32(Functional, FunctionalFloat32Only, PytorchTestCase):
     dtype = torch.float32
     device = torch.device('cpu')
