--- test/torchaudio_unittest/rnnt/autograd_cuda_test.py
@@ -1,11 +1,0 @@
-import torch
-from .autograd_impl import Autograd
-from torchaudio_unittest import common_utils
-from .utils import skipIfNoRNNT
-
-
-@skipIfNoRNNT
-@common_utils.skipIfNoCuda
-class TestAutograd(Autograd, common_utils.PytorchTestCase):
-    dtype = torch.float32
-    device = torch.device('cuda')
