--- test/torchaudio_unittest/rnnt/torchscript_consistency_cpu_test.py
@@ -1,10 +1,0 @@
-import torch
-
-from torchaudio_unittest.common_utils import PytorchTestCase
-from .utils import skipIfNoRNNT
-from .torchscript_consistency_impl import RNNTLossTorchscript
-
-
-@skipIfNoRNNT
-class TestRNNTLoss(RNNTLossTorchscript, PytorchTestCase):
-    device = torch.device('cpu')
