--- test/torchaudio_unittest/rnnt/torchscript_consistency_cuda_test.py
@@ -1,11 +1,0 @@
-import torch
-
-from torchaudio_unittest.common_utils import PytorchTestCase, skipIfNoCuda
-from .utils import skipIfNoRNNT
-from .torchscript_consistency_impl import RNNTLossTorchscript
-
-
-@skipIfNoRNNT
-@skipIfNoCuda
-class TestRNNTLoss(RNNTLossTorchscript, PytorchTestCase):
-    device = torch.device('cuda')
