+++ test/torchaudio_unittest/transforms/torchscript_consistency_cpu_test.py
@@ -1,10 +1,10 @@
 import torch
 
 from torchaudio_unittest.common_utils import PytorchTestCase
-from .torchscript_consistency_impl import Transforms
+from .torchscript_consistency_impl import Transforms, TransformsFloat32Only
 
 
-class TestTransformsFloat32(Transforms, PytorchTestCase):
+class TestTransformsFloat32(Transforms, TransformsFloat32Only, PytorchTestCase):
     dtype = torch.float32
     device = torch.device('cpu')
