+++ test/torchaudio_unittest/transforms/torchscript_consistency_cuda_test.py
@@ -1,11 +1,11 @@
 import torch
 
 from torchaudio_unittest.common_utils import skipIfNoCuda, PytorchTestCase
-from .torchscript_consistency_impl import Transforms
+from .torchscript_consistency_impl import Transforms, TransformsFloat32Only
 
 
 @skipIfNoCuda
-class TestTransformsFloat32(Transforms, PytorchTestCase):
+class TestTransformsFloat32(Transforms, TransformsFloat32Only, PytorchTestCase):
     dtype = torch.float32
     device = torch.device('cuda')
