+++ torchaudio/functional/__init__.py
@@ -25,6 +25,7 @@ from .functional import (
     resample,
     edit_distance,
     pitch_shift,
+    rnnt_loss,
 )
 from .filtering import (
     allpass_biquad,
@@ -98,4 +99,5 @@ __all__ = [
     'resample',
     'edit_distance',
     'pitch_shift',
+    'rnnt_loss',
 ]
